distrodefs: rename "install.iso" to "installer.iso" for bootc
Our installers are usually called "installer.iso". For some
(probably silly) reason they are called "install.iso" in
the bootc distro definitions. This is not a big deal as
image-builder will rename the artifcat usually into a
more meaningful name anyway but it seems nice to change
this for conistency.