osbuild: introduce FileChanger interface and use it
It is useful for the higher layers to know what files are
touched by a specific stage. This is used in e.g. the
bootc disk where we want to target the relabeling.

So this commit introduce a new `FileChanger` interface
that can be implemented by the relevant stages and is
then used in RawBootcImage to relabel only the specific
paths that the stages touch.

Build on top of the excellent work by Alex in PR#2037.