manifest: Add PXE support to the raw bootc pipeline
In order to PXE boot a compressed bootc filesystem we need to run dracut
to add the dmsquash-live module to the initramfs. It also needs to
disable composefs (it doesn't work when booting from a squashfs
compressed root filesystem with dmsquash-live).

This also creates an empty /usr directory in the root of the filesystem,
this is needed so that dmsquash-live will recognize the compressed
filesystem as a rootfs. Ideally dmsquash-live should be patched to also
look for the /ostree directory.

When booting bootc using PXE and a compressed filesystem it doesn't make
sense to run growpart, or to run the nightly update timer (with no
persistent storage it will just use up ram until rebooted). So this also
masks:
    bootc-generic-growpart.service
    bootc-fetch-apply-updates.timer

Related: HMS-9767