test: skip build/boot tests as non-root
This skips the build and boot tests for non-root users, they
are currently required to run as root. A nice side effect is
that the normal workflow of running pytest to run the "normal"
tests is not interrupted. We could of course also use pytest
and a custom "marker" like "@pytest.mark.integration" if
this commit is considered too loose.