manifest: relabel /etc and /var after bootc install
In the RawBootcImage struct, ff we add stages after
`bootc install to-filesystem` we need to relabel the changed
files.

Initially we did this in a very coarse matter, i.e. just relabeld
the entire filesystem. This can lead to bugs like issue#1161 so
Alex fixed this by adding fine grained relabeling in PR#2037.

However now the "manifest" packages has to deal with many
details of the "osbuild" package which ideally we would like
to keep hidden. So this commit changes the approach again
to be less fine grained and just relabels "/etc/" and "/var"
if we add any stages after "boot install to-filesystem". This
should be sufficient as we know its a bootc system so the
only mutable content is in /etc and /var and we have files/dir
policies that prevent modifications outside for blueprint
content.

This is a (much) simpler alterantive to PR#2042 which also
address this area but keeps the fine grained control of
the original work.