pkg/rpmmd: Remove `BaseURL` from `Package` struct
While working on the osbuild-depsolve-dnf parts, I realized that the
`BaseURL` information in this case comes from the repository metadata,
where it is almost never set. So in reality, the value would be almost
always empty. The baseurl value is and should be deduced from the
repository config that the package comes from.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>