manifest: add ISOTree for generic ISOs
Our other ISOTree is tighly-coupled to anaconda. However, we are trying
to build a fully generic ISO built from a bootc container image, so we
cannot make any assumption about it having anaconda. This new pipeline
is exactly meant for this. We should probably try to deduplicate it in
the future, but let's keep it separate as it makes experimentation
simpler.