image: add container_based_iso
This is meant as a generic ISO built from a bootc container. The goal
here is that the ISO should be fully customizable in a Containerfile (or
any container image build tooling). We are not there yet: e.g. we
hardcode the kernel args, and we assume that this is a live ISO. I hope
that following PRs/commits will be able to fix that.