bootc: add bootc-generic-iso
Unlike bootc-installer, this is a fully "payload-agnostic" ISO. Most
importantly, it doesn't assume there's anaconda inside. Moreover, it
currently makes an assumption that the user wants a live ISO. We should
make this configurable in the future (together with customizing kernel
args). The goal is that users can just do all their customizations in a
Containerfile, and then just pass this container image to image-builder
and the bootc-generic-iso image type, and it will just build a working
ISO.