fedora: write `/etc/sysconfig/kernel`
When the `/etc/sysconfig/kernel` file does not exist with the
appropriate values then the default kernel does not get updated when a
new kernel is installed with DNF.

Normally this file is written by either Kiwi [1] or Anaconda [2] hence
we should write it as well. We already wrote this for RHEL-likes.

[1]: https://pagure.io/fedora-kiwi-descriptions/blob/rawhide/f/root/etc/sysconfig/kernel
[2]: https://github.com/rhinstaller/anaconda/blob/8ea01bacf57df26dcc14d736218bc02bbb07f766/pyanaconda/modules/storage/bootloader/utils.py#L148

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>