gen-manifests: add -print-only
We need a `-print-only` option so that we can see what
gen-manifests would generate. This will be input for a pytest
based framework (but seems generally be useful).

This also tweaks the outputs so that any status message now
goes to stderr so that one can parse stdout for the list
of image manifests that would be generated.