pkg/depsolvednf: don't return `rhsmMap` from makeDepsolveRequest()
The `rhsmMap` was previously constructed in `makeDepsolveRequest()`,
but not used at all and returned to the caller. The caller then used it
to interpret the Solver API response. Considering this, the
makeDepsolveRequest() is simply the wrong place to construct `rhsmMap`.

Move the construction of the `rhsmMap` to the caller. This decouples the
request construction even more from the Solver implementation. This will
make decoupling the wire-format construction from the Solver
implementation much cleaner.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>