pkg/depsolvednf: add v1 API apiHandler interface implementation
Add a version of the current (v1) API implementation, which implements
the apiHandler interface. The implementation is not used yet. It will be
swapped with the current implementation that is coupled with the Solver,
in the following commits.

The goal is to keep the API as is, without any noticeable change.

Add unit tests for make*Request() methods, verifying that the produced
JSON is as expected. The parse*Result() will be tested indirectly by
the unit tests that verify the Solver methods, when they switch to use
API handler implementation.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>