pkg/depsolvednf: replace Request.Hash() with hashRequest() helper
The caching for the dump and search commands depended on the API request
wire-format struct. To not need to complicate the API handler interface
with this specific aspect, replace the old way of calculating hash by
hashing the whole raw request. Although the new implementation will
produce different hashes than the old one, this is a one time cache
invalidation. The functionality is technically the same.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>