Revert "test: add fallbacks for openscap datastreams"
This reverts commit 632078941ee9847b94389b0eb42cc1ec36ec4937.

This commit assuemd that /usr/share/xml/scap/ssg/content/
has only the matching profiles installed but on some
systems (rhel8) that is not the case and there we have
multiple ones. So revert the incorrect commit for now and
later add code that /etc/os-release to find the expected
one.