scripts/build-image: use libosinfo to detect installer
When using virt-install --location disk.iso the libosinfo library is
used to detect operating system and path to kernel and initramdisk on
the ISO image. This is done by comparing disk label against its database
of known OSes.

While most of the images I tested do work, to fully ensure this is the
case and to prevent regressions, let's run this tool after image is
built in case it is an ISO.

For unknown OS versions, the database contains "unknown" entries which
are still recognized as unreleased versions, so this should work even if
our runners will have older version of the library. This works for RHEL,
but not for Fedora where we must use "dvd" suffix. It also does not work
for non-x86_64 architectures so added a workaround until this is fixed
in the libosdb.