osinfo: use more paths to search uefi vendor on bootc images
We got a bugreport that the anaconda-iso bootc image type
does no longer work on f44. It turns out this is because
`/usr/lib/bootupd/updates/` does no longer contain a copy
of the data. So as a pragmatic solution use
`/usr/lib/efi/shim/*/EFI/*` as an additional search path.

Note that ideally we would have some way to querry bootupd
or the EFI.json but it seem no interface for this is available
and also that the EFI.json does not contain info about shim-x64
but only about the (not installed) shim package on f44.

Closes: https://issues.redhat.com/browse/HMS-9867