test: revert pytest part of `generate-ostree-build-config`
In https://github.com/osbuild/images/pull/2045/ the script
was converted to use pytest. However this does not work
for ostree configs. A ostree config must be rewritten so
that it points to an actual ostree commit and server.

This is done in the `setup_dependencies()` script that then
creates a different config/config_name. As this is different
from the configs we get via `gen-manifests --dry-run` pytest
cannot be used here.

So for now be pragmatic and just revert
https://github.com/osbuild/images/pull/2045/files#diff-8a0342654218957707e9769f498c86df3f070ee601f5b059f77b30d4d1f50d1eL18

We probably want get back to this eventually and make this
setup part of pytest.