test: run cross arch smoke test on ppc64le, s390x
This commit adds a minimal smoke test for centos-10/qcow2 that
will create a qcow2 and boot it. It demos how we could do more
cross arch testing and keeps the qemu code tested.

We need to decide if we want to do more tests like this, its
a trade-off. OTOH its nice to have (some) assurance that our
images boot (we did break s390/ppc64 partition tables in the
past by accident). OTOH its a bit of a pain when something fails
to figure out of it is failing because of some qemu incompatibilites
or because there is a real issue in the image.

Having this at least for local testing/validation is probably
useful though.

Note that the s390x test needs qemu-user >= 7.2-rc2 so this will
not run currently (I did run it on my local machine and it works
but the kdump service takes > 5min to startup).