test: add `--keep-booted` option to `boot-image`
Add a convenience way to keep the booted image around when
using `test/scripts/boot-image`. When running:
```console
$ ./test/scripts/boot-image --keep-booted <path>
```
the image will stay around after the check script was run
so that the image can be inspected. It also provides a
convenient output how to login into the system:
```
$ ./test/scripts/boot-image --keep-booted ./build/centos_10-x86_64-qcow2-empty/
...
***********************************
keeping the image ./build/centos_10-x86_64-qcow2-empty/qcow2/disk.qcow2 booted as requested, press enter or ctrl-c to stop
to connect run:
ssh -i /tmp/tmpxbfzwzlo/testkey -p 50635 -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no osbuild@localhost
```

Note that currently only qcow2 is supported but if this is
considered useful we can expand it to the other image types.

For some (like ISO image-installer) this is tricky as we
cannot inject credentials via cloud-init. But that is a
more general problem that we would ideally tackle at some
point.