bootc: Fix selinux labeling when using separate build container
When the build container and the target container are using different
selinux policies we need to label them as such. The build container
needs to be labeled with img.OSCustomizations.BuildSELinux,
and the "target" container should be labeled with the same policy
as the source target.

Currently we're using the build policy for both, and in the case where
the source container doesn't include that policy it fails.  For
example, if the build container is using the targeted policy, but the
bootc source image only has the automotive policy, then the relabeling
in the target pipeline atm fails with "targeted policy not found".