manifest: add `set -e` to `bootc switch...` kickstart %post
The `bootc switch` command we use in our bootc installer ISOs
stopped working [0] with the error:
```
error: Switching: Initializing storage: System not booted via bootc
```

However the install continue because while we do set
`--erroronfail` the script is not setting `set -e`.

This commit fixes this and makes the bootc switch error in script
a hard failure. This means that installs will error currently
but its better to error than to install a system that is then
not connected to any updates :(

[0] https://github.com/osbuild/image-builder-cli/issues/410