many: copy files into ISO
Allow defining files that need to be copied to the root of an ISO. For
example current Fedora ISOs container a `Fedora-Legal-README.txt` and a
`LICENSE` file that we are not including.

This commit allows defining files to be copied, subsequent commits
implement copying and add the copying of files to Fedora's definitions.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>