distro: plumb in iso config
Create a new configuration struct that determines options for ISO media.
This follows the same idea as we use for `image_config` (which likely
should be split out more) and `installer_config`.

Also introduce the function that will in the future handle merging of
the `iso_config` and any customization values and apply this function to
each image that includes `ISOCustomizations`.

Note that nothing here is really used yet still; this is only setup code
to be able to use the `ISOCustomizations` in the next commits.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>