vmtest: disk image format fixed
Cross arch testing does not work properly, the QEMU command line renders
image twice. Once as device option and once as an argument. On top of
that, arguments are incorrectly formatted since Path object is getting
printed:

PosixPath('/var/tmp/tmpdk5gg6b6/netinst.iso'), PosixPath('/var/tmp/tmpp40bxujp/disk.img')

QEMU complains about raw image that cannot be auto-detected because it
is empty. The problem is, for installer ISO we create an empty raw image
and since there is zero content it cannot be autodetected. But the
format= argument is passed as qcow2 which is not correct in this case.