vmtest: use x86_64-v3 instead of host CPU passing
Passing cpu=host creates inconsistent environment when running on
various laptops locally or even in AWS when running on different
instance types. It is better to be explicit to always use the same CPU
feature set.

This patch uses Skylake family which is x86_64-v3 which is required by
RHEL10. It also adds few extra features for better performance. Finally,
Q35 is used instead of the legacy platform.