vmtest: use all available cores
We hardcode the number of cores to use in the QEMU command line. This is not
ideal, as it makes it harder to test with different numbers of cores.

Instead, we should use the `-smp` argument to QEMU to specify the number of
cores to use. This is more flexible and allows us to test with different
numbers of cores.